feat:implement mDNS bridge service discovery feat:add tokio runtime for tauri
this commit implements the primary way of discovering hue bridge
services on the local network as described in [Hue Bridge Discovery](https://developers.meethue.com/develop/application-design-guidance/hue-bridge-discovery/).

this commit also makes the necessary changes to use the same tokio
runtime for the application code as for the tauri core.

this commit also adds a logging framework and provides a default
configuration for that.