feat:add `clap` argument parser ops:generate typescript bindings before prod build
this commit introduces the `clap` crate - the command line argument
parser library. with it we can now run the tauri backend with the
`--generate-bindings-only` flag to only output the typescript bindings
and then exit the program. this, we can now use to generate typescript
bindings freshly before every production build. i have also went ahead
and added the generated `src/bindings.ts` file to `.gitignore` so as not
to track it and forcing people to generate it anew on every build.